requirements.txt
google-adk
google-genai
google-cloud-aiplatform[adk,agent_engines]
google-cloud-geminidataanalytics
pandas
